refactor: extract utils from tracing.ts to new utils.ts file
